Refactored code and updated PPP
