+++ docs/DeveloperGuide.adoc
@@ -124,6 +124,8 @@ image::DeleteSequenceDiagram.png[align="center"]
 
 NOTE: The lifeline for `StudentRemoveCommandParser` should end at the destroy marker (X) but due to a limitation of PlantUML, the lifeline reaches the end of diagram.
 
+// tag::design-model[]
+
 [[Design-Model]]
 === Model Component
 .Structure of the Model Component
@@ -182,7 +184,7 @@ The `Module`,
 
 *API* : link:{repoURL}/src/main/java/seedu/planner/commons/util/ModuleDataImporterUtil.java[`ModuleDataImporterUtil.java`]
 
-The `ModuleDataImportUtil`,
+The `ModuleDataImporterUtil`,
 
 * is a helper class used to load Modules from a JSON file downloaded from NUSMods API
 * handles the conversion of Modules from JSON format to `Module` objects using `JsonSerializableModule` and `JsonAdaptedModule`
@@ -241,6 +243,9 @@ The `SpecialisationGraduationRequirement` class,
 * contains a list of `GraduationRequirement` objects to represent a Specialisation's elective modules
 * handles the calculation of fulfillment for the `SpecialisationGraduationRequirement` object based on `Major` and primary and elective modules
 
+// end::design-model[]
+
+// tag::design-storage[]
 
 [[Design-Storage]]
 === Storage Component
@@ -256,6 +261,10 @@ The `Storage` component,
 * can save the Planner data in json format and read it back.
 * consists of various JSON implementations of Models
 * handles the saving and loading of Models to provide stateful behaviour in the application
+
+Additional details on how certain Models were stored in JSON format is available at <<Backend-JSON>>.
+
+// end::design-storage[]
 
 [[Design-Commons]]
 === Common Classes
@@ -396,7 +405,7 @@ This implementation consist of 3 commands.
 * `LessonRemove` - Remove the lessons from the active student based on the provided index.
 
 ===== Adding of Lesson
-The `LessonAddCommandParser` method is called and it will get the `ModuleCode` and `Semester` provided by the arguments. It will run the `LessonDataImporter` class run method to convert the json file to `Lesson` Object. This is then added to the active student on the planner.
+The `LessonAddCommandParser` method is called and it will get the `ModuleCode` and `Semester` provided by the arguments. It will run the `LessonDataImporterUtil` class run method to convert the json file to `Lesson` Object. This is then added to the active student on the planner.
 
 ===== Listing of Lesson
 The `LessonListCommand` method is called and it will iterate through the list of `Lesson` that is associated with the active student in the planner.
@@ -673,15 +682,60 @@ Given below is an example usage scenario and how the `list` mechanism behaves at
 
 Certain properties of the application can be controlled (e.g user prefs file location, logging level) through the configuration file (default: `config.json`).
 
+// tag::backend-json[]
 [[Backend-JSON]]
-=== JSON Handling & Live Data Management
+=== JSON Handling & Module Data Management
 
 As mentioned in several sections (of <<Implementation>>) above, NUS Module Planner has the capability to pull live data for entities, such as, `Major` and `Module`, via *JSON*.
 
-*JSON* stands for _JavaScript Object Notation_ and is a lightweight format for storing and transporting data. We choose this syntax as it is "self-describing" and thus, easy to understand, while keeping our application size low.
+*JSON* stands for _JavaScript Object Notation_ and is a lightweight format for storing and transporting data. We choose this syntax as it is "self-describing" and thus, human-readable and easy to understand, while keeping our application size low.
 
-_We will add more about how we implemented JSON and how we get, format, read and store the data._
+*Jackson* is a standard JSON library for Java that contains a suite of data-processing tools, including JSON parsers and generators. Jackson is used for the JSON-related operations mentioned below.
 
+The original schema of the JSON file came from NUSMods via version 2 of their API available link:http://nusmods.com/api/v2[here]. The endpoints we used were:
+
+* `/{acadYear}/moduleInfo.json`
+* `/{acadYear}/modules/{moduleCode}.json`
+
+The Academic Year (`acadYear`) used is: `2019/2020`.
+
+From the JSON responses of the above API endpoints, we created several classes to help convert the JSON representation of the object into a working Java Object that the application can use, as well as the other direction which is used to stored the current state of the `Planner` in a JSON file.
+
+Such classes include:
+
+* `JsonSerializableLesson`
+* `JsonSerializableModule`
+* `JsonSerializablePlanner`
+* `JsonSerializableSemesterData`
+
+The `ModuleDataImporterUtil` is used to parsed the raw JSON files containing Module information and convert them into `JsonSerializableModule` objects which contain various attributes such as a list of `JsonSerializableSemesterData`.
+
+.Sequence Diagram for a `run` operation in `ModuleDataImporterUtil`
+image::sequenceDiagramModuleDataImporterUtil.png[align="center"]
+
+The `LessonDataImporter` is used to parsed the raw JSON files containing Module-specific Lesson information and convert them into `JsonSerializableModule` objects which contain various attributes such as a list of `JsonSerializableLesson`.
+
+.Sequence Diagram for a `run` operation in `LessonDataImporter`
+image::sequenceDiagramLessonDataImporterUtil.png[align="center"]
+
+The following classes are used to convert objects of various classes (such as Module, Lesson, Student) into Jackson-friendly versions.
+
+* `JsonAdaptedEnrollment`
+* `JsonAdaptedGrade`
+* `JsonAdaptedLesson`
+* `JsonAdaptedModule`
+* `JsonAdaptedModuleCode`
+* `JsonAdaptedSemesterData`
+* `JsonAdaptedSemesterYear`
+* `JsonAdaptedStudent`
+* `JsonAdaptedStudentSemester`
+* `JsonAdaptedTimeTable`
+* `JsonAdaptedTimeTableMap`
+* `JsonAdaptedTimeTablePair`
+
+When saving the `Planner`, `JsonSerializablePlanner` is used to stored all the various attributes of the `Planner` in a JSON-compatible format, such as list of `Student`, currently active `Student` index, currently active `SemesterYear`.
+
+// end::backend-json[]
 === Logging
 
 We are using `java.util.logging` package for logging. The `LogsCenter` class is used to manage the logging levels and logging destinations.
@@ -714,6 +768,8 @@ Refer to the guide <<Testing#, here>>.
 
 Refer to the guide <<DevOps#, here>>.
 
+// tag::proposed-future-enhancements[]
+
 == Proposed Future Enhancements
 
 NUS Module Planner has a lot more potential to grow!
@@ -726,6 +782,12 @@ Below are some of the other great features we think can be implemented in v2.0 o
 * *Requirements Double-Upper* - Instead of just one `Major`, handle double `Major` programme students, or even those with Minors.
 * *Module Popularity Prediction* - Get and analyse historic data from NUSMods about module uptake, so that students do not have to plan for modules that would be overly subscribed.
 * *Support for Overseas Exchange and University Town College Programmes* - Be able to plan for special modules only offered in these programmes, if a student is enrolled in them.
+* *Live Module Data* - Get the latest Module Data available from NUS Mods and plan your Semester before it starts.
+* *Module Data History* - View historical data on Module enrollments and decide if when you want to take a certain `Module`.
+
+// end::proposed-future-enhancements[]
+
+// tag::appendix-product-scope[]
 
 [appendix]
 == Product Scope
@@ -743,6 +805,8 @@ Below are some of the other great features we think can be implemented in v2.0 o
 
 * aggregates all information regarding modules and graduation requirements in a single app
 * manages academic progress and module planning faster than a typical mouse/GUI driven app
+
+// end::appendix-product-scope[]
 
 [appendix]
 == User Stories
