+++ src/main/java/seedu/planner/storage/JsonAdaptedSemesterData.java
@@ -1,4 +1,4 @@
-package seedu.planner.model.module;
+package seedu.planner.storage;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -7,6 +7,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.planner.commons.exceptions.IllegalValueException;
+import seedu.planner.model.module.SemesterData;
 
 /**
  * Jackson-friendly version of {@link SemesterData}.
